refactor: use `devtools` as command rather than `enable`
