+++ packages/nuxi/src/commands/index.ts
@@ -18,8 +18,7 @@ export const commands = {
   info: () => import('./info').then(_rDefault),
   init: () => import('./init').then(_rDefault),
   create: () => import('./init').then(_rDefault),
-  enable: () => import('./enable').then(_rDefault),
-  disable: () => import('./disable').then(_rDefault),
+  devtools: () => import('./devtools').then(_rDefault),
   upgrade: () => import('./upgrade').then(_rDefault),
   test: () => import('./test').then(_rDefault),
   add: () => import('./add').then(_rDefault),
