--- packages/nuxi/src/utils/features.ts
@@ -1,7 +1,0 @@
-export const featureCLIs = {
-  devtools: '@nuxt/devtools@latest'
-} as const
-
-export function isFeature (feature: string): feature is keyof typeof featureCLIs {
-  return feature in featureCLIs
-}
